+++ NEWS.md
@@ -1,5 +1,14 @@
 # kwb.pkgbuild 0.1.0.9000
 
+* Improved documentation generation (closes #6)
+
+   +  changed `README.Rmd` to `index.Rmd` (used by `pkgdown::build_home`) and added 
+      "Installation" chapter 
+
+   + added "Installation" chapter for 
+
+* automatically convert package title to Title Case with `tools::toTitleCase()` 
+
 * Added a `NEWS.md` file to track changes to the package.
 
 * see http://style.tidyverse.org/news.html for writing a good `NEWS.md`